+++ Source/astcenc_kmeans_partitioning.cpp
@@ -258,7 +258,7 @@ static void basic_kmeans_update(
 
 	for (int i = 0; i < partition_count; i++)
 	{
-		cluster_centers[i] = color_sum[i] * (1.0f / weight_sum[i]);
+		cluster_centers[i] = color_sum[i] * (1.0f / static_cast<float>(weight_sum[i]));
 	}
 }
